Attempt to fix Issue # 6  Multiple forms.
Form opening is controlled by method7() in Mainform.cs.   It searches all child forms to see if a form is already open and activates it if it is.  Is seems to be finding the open form but falling through and opening a new one anyway.   Not totally sure I understand why the code has changed or needs changing if it worked before but this change seems to fix it.  Have a close look and see if you can make more sense of it than me.